% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxiety-data.R
\docType{data}
\name{anxiety}
\alias{anxiety}
\title{Anxiety Dataset}
\format{A dataframe of 1000 observations and 3 variables.
\describe{
  \item{math}{math scores}
  \item{english}{english scores}
  \item{anxiety}{anxiety level (low, med, high)}
}}
\source{
\href{https://statistics.laerd.com/stata-tutorials/one-way-manova-using-stata.php}
{Laerd Statistics}
}
\usage{
data(anxiety)
}
\description{
An individual's math and english scores based on their anxiety level. 
Dataset is useful for running a One-Way Manova. Dataset is 
simulated but draws inspiration from Laerd Statistics.
}
\keyword{datasets}
