% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleEFA-data.R
\docType{data}
\name{ExampleEFA}
\alias{ExampleEFA}
\title{Example EFA dataset}
\format{A dataframe of 1365 observations and 13 variables.
\describe{
  \item{X}{Unique ID}
  \item{Item 13}{Instructor well prepared}
  \item{Item 14}{Instructor scholarly grasp}
  \item{Item 15}{Instructor confidence}
  \item{Item 16}{Instructor focus lectures}
  \item{Item 17}{Instructor uses clear relevant examples}
  \item{Item 18}{Instructor sensitive to students}
  \item{Item 19}{Instructor allows me to answer questions}
  \item{Item 20}{Instructor accessible to students outside class}
  \item{Item 21}{Instructor aware of students understanding}
  \item{Item 22}{I am satisfied with student performance evaluation}
  \item{Item 23}{Compared to other instructors, this instructor is}
  \item{Item 23}{Compared to other courses this course was}
}}
\source{
\href{https://stats.idre.ucla.edu/spss/output/factor-analysis/}
{Institute for Digital Research and Education}
}
\usage{
data(ExampleEFA)
}
\description{
Student responses on course evaluations. 
Adapted from a dataset from UCLA's
Institute for Digital Research and Education.
}
\keyword{datasets}
