% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{initialize_backpack}
\alias{initialize_backpack}
\title{Initalize Backpack}
\usage{
initialize_user_views()
}
\arguments{
\item{path}{file path to store the table with user-defined binders in. Creates an environment variable where the path is stored.}
}
\description{
Function to initialize the file where user-defined binders will be saved.
}
\examples{
initialize_backpack("~/Documents/") ## to store the user binders in documents folder. 

}
