% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{install_binders}
\alias{install_binders}
\title{Install Binders}
\usage{
install_binders(binders)
}
\arguments{
\item{binders}{The binder to install}
}
\description{
Function to install the packages in the binders.
}
\examples{
install_binders("MLM")

install_binders()
}
