% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{list.base.views}
\alias{list.base.views}
\title{List Views available}
\usage{
list.base.views()
}
\description{
This function allows you to see the list of views available.
}
\keyword{list.base.views()}
\keyword{views}
