% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{list_binders}
\alias{list_binders}
\title{List Binders}
\usage{
list_binders()
}
\arguments{
\item{compartment}{The compartment to view the binders in. Defaults to "all". 
- "all" to view all the binders
- "master" to view binders that came with the package
- "user" to view binders created by the user}
}
\description{
Function to see the binders of packages available and a summary for each.
}
\examples{
list_binders(compartment="master")

list_binders()
}
