% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessons.R
\name{list_lessons}
\alias{list_lessons}
\title{List Lessons}
\usage{
list_lessons(name)
}
\arguments{
\item{name}{Name of the lesson to learn.}
}
\description{
Function to list available lessons.
}
\examples{
list_lesson()

}
