% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{load_binders}
\alias{load_binders}
\title{Load Binders}
\usage{
load_binders(binders)
}
\arguments{
\item{binders}{character vector of binder names. The binder(s) whose packages are to be loaded into the environment.}
}
\description{
Function to load the packages in a binder(s).
}
\examples{
load_binders("MLM")

}
