% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_ed-data.R
\docType{data}
\name{psych_ed}
\alias{psych_ed}
\title{Psych_Ed Dataset}
\format{A dataframe of 300 observations of 7 variables.
\describe{
  \item{control}{locus of control score}
  \item{concept}{self-concept score}
  \item{motivation}{motivation score}
  \item{read}{standardized test score in reading}
  \item{write}{standardized test score in writing}
  \item{math}{standardized test score in math}
  \item{science}{standardized test score in science}
}}
\source{
\href{https://stats.idre.ucla.edu/stat/data/mmreg.csv}{UCLA Institute for Digital Research and Education}
}
\usage{
data(psych_ed)
}
\description{
Measurements of psychological constructs and scores on educational tests. Dataset adapted from dataset
 found on website for UCLA's Institute for Digital Research and Education. See reference link below for
  original.
}
\keyword{datasets}
