% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seeds}
\alias{seeds}
\title{Seeds Dataset}
\format{An object of class \code{data.frame} with 210 rows and 8 columns.}
\source{
\href{http://archive.ics.uci.edu/ml/datasets/seeds?ref=datanews.io}
{UCI Machine Learning Repository}
}
\usage{
data(seeds)
}
\description{
Measurements of geometrical properties of kernels belonging to three 
different varieties of wheat. Adapted from the dataset taken from the 
UCI Machine Learning Repository.
}
\keyword{datasets}
