% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{set_backpack_path}
\alias{set_backpack_path}
\title{Set backpack path}
\usage{
set_backpack_path()
}
\arguments{
\item{path}{file path where user-defined binders are. Creates an environment variable where the path is stored.}
}
\description{
creates the environment variable "PATH_TO_BACKPACK_USER_VIEWS" where the backpack path will be stored. This is referenced internally to look for user's personla binders.
}
\examples{
set_backpack_path("~/Documents/") 

}
