% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessons.R
\name{start_lesson}
\alias{start_lesson}
\title{Start Lesson}
\usage{
start_lesson(lesson_name)
}
\arguments{
\item{lesson_name}{Name of the lesson to learn.}
}
\description{
Function to start a lesson.
}
\details{
This can also be done using the learnr package as follows 
'learnr::run_tutorial("2004-MatrixAlgebra",package="backpack")'
}
\examples{
start_lesson(name="2004-MatrixAlgebra")

start_lesson()
}
