% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_packages.R
\name{unbind_packages}
\alias{unbind_packages}
\title{Unbind Packages}
\usage{
unbind_packages(Package, binder)
}
\arguments{
\item{package}{: package(s) to remove. Pass a vector if multiple packages.}

\item{binder}{: specifies the binder from which Packages need to be removed. If Packages is NULL, will remove entire binder.}
}
\description{
Function to remove packages from existing binders, or entire binders, from user views.
}
\examples{
unbind_packages(Package = c("a","b"), binder = "alphabets") 
unbind_packages(binder = 'alphabets')



}
