% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{uninstall_binders}
\alias{uninstall_binders}
\title{Uninstall Binders}
\usage{
uninstall_binders(binders)
}
\arguments{
\item{binders}{The binder(s) to uninstall}
}
\description{
Function to uninstall the packages in the binder(s).
}
\examples{
uninstall_binders("MLM")

uninstall_binders()
}
