% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseViewsFunctions.R
\name{view_binders}
\alias{view_binders}
\title{View Binders}
\usage{
view_binders()
}
\arguments{
\item{compartment}{character vector, the compartment to view the binders in. Defaults to "all". 
- "all" to view all the binders
- "master" to view binders that came with the package
- "user" to view binders created by the user}

\item{search}{character string, regex style search string for binders with matches to words in the search string. Deafults to NULL in which case all the binders are shown.}
}
\description{
Function to see the binders of packages available and a summary for each.
}
\examples{
view_binders(compartment="master") ## to view all the binders in the master list
view_binders(compartment="master",binders="Machine Learning") ## to view a summary of just the Machine Learning Binder

}
