% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_packages.R
\name{bind_packages}
\alias{bind_packages}
\title{Bind Packages}
\usage{
bind_packages(Package, binder, source)
}
\arguments{
\item{Package}{: new packages to add}

\item{binder}{: either an existing binder or a new user-defined binder}

\item{source}{: whether package source is CRAN or Github Repo}
}
\description{
Function to add packages to existing binders to the user views, or to create
new binders to store packages
}
\examples{
bind_packages(Package = c("a","b","c"),binder = "alphabets", source = "CRAN")

bind_packages()
}
