% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds-data.R
\docType{data}
\name{seeds}
\alias{seeds}
\title{Seeds Dataset}
\format{A dataframe of 210 observations of 8 variables.
\describe{
  \item{area}{area of kernel}
  \item{perimeter}{perimeter of kernel}
  \item{compact}{compactness of kernel given by C = 4*pi*A/P^2}
  \item{length}{length of kernel}
  \item{width}{width of kernel}
  \item{asymm}{asymmetry coefficient}
  \item{groove}{length of kernel groove}
  \item{variety}{variety of wheat: 1 = Kama, 2 = Rosa and 3 = Canadian}
}}
\source{
\href{http://archive.ics.uci.edu/ml/datasets/seeds?ref=datanews.io}{UCI Machine Learning Repository}
}
\usage{
data(seeds)
}
\description{
Measurements of geometrical properties of kernels belonging to three 
different varieties of wheat. Adapted from the dataset taken from the 
UCI Machine Learning Repository.
}
\keyword{datasets}
